requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.2.1
coverage==7.5.1
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==42.0.5
cycler==0.12.1
daphne==4.0.0
Django==4.2
django-browser-reload==1.12.0
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-restframework==0.0.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
flower==2.0.1
fonttools==4.51.0
html2pdf==0.1
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kiwisolver==1.4.5
kombu==5.3.5
matplotlib==3.8.4
msgpack==1.0.7
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pyparsing==3.1.2
pytest==8.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
seaborn==0.13.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tomli==2.0.1
tornado==6.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
